requirements.txt
azure.identity
azure.storage.blob
azure.iot.device